--- a/cdc/model/sink.go
+++ b/cdc/model/sink.go
@@ -24,6 +24,7 @@ type RowChangedEvent struct {
 	// IndieMarkCol will be set to the name of the unique index
 	IndieMarkCol string
 	Columns      map[string]*Column
+	Keys         []string
 }
 
 // ToMqMessage transforms to message key and value