--- a/cdc/sink/causality.go
+++ b/cdc/sink/causality.go
@@ -5,41 +5,45 @@ package sink
 // if some conflicts exist in more than one groups, then syncer waits all SQLs that are grouped be executed and reset causality.
 // this mechanism meets quiescent consistency to ensure correctness.
 type causality struct {
-	relations map[string]struct{}
+	relations map[string]int
 }
 
 func newCausality() *causality {
 	return &causality{
-		relations: make(map[string]struct{}),
+		relations: make(map[string]int),
 	}
 }
 
-func (c *causality) add(keys []string) error {
+func (c *causality) add(keys []string, idx int) {
 	if len(keys) == 0 {
-		return nil
+		return
 	}
 
 	for _, key := range keys {
-		c.relations[key] = struct{}{}
+		c.relations[key] = idx
 	}
-	return nil
 }
 
 func (c *causality) reset() {
-	c.relations = make(map[string]struct{})
+	c.relations = make(map[string]int)
 }
 
 // detectConflict detects whether there is a conflict
-func (c *causality) detectConflict(keys []string) bool {
+func (c *causality) detectConflict(keys []string) (bool, int) {
 	if len(keys) == 0 {
-		return false
+		return false, 0
 	}
 
+	firstIdx := -1
 	for _, key := range keys {
-		if _, ok := c.relations[key]; ok {
-			return true
+		if idx, ok := c.relations[key]; ok {
+			if firstIdx == -1 {
+				firstIdx = idx
+			} else if firstIdx != idx {
+				return true, -1
+			}
 		}
 	}
 
-	return false
+	return firstIdx != -1, firstIdx
 }