<!--
Thank you for contributing to TiDB-CDC! Please read MD's [CONTRIBUTING](https://github.com/pingcap/tidb-cdc/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Close https://github.com/pingcap/ticdc/issues/463.

bench test:

sink 1 table, 10000 rows to mysql

before this PR: 59s

This PR: 
```shell
8.5s.  -- 16 workers.
10.5s.  -- 8 workers.
15.65s -- 4 workers.
```

### What is changed and how it works?

Most same with the [DM](https://github.com/pingcap/dm), use [causality](https://github.com/pingcap/dm/blob/master/syncer/causality.go) to do conflict check, change a little.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test, in adding.

Code changes

 - Has exported function/method change

Side effects

 - Possible performance regression
 - Increased code complexity
 - Breaking backward compatibility

Related changes

 - Need to cherry-pick to the release branch
 - Need to update the documentation
 - Need to update the `tidb-cdc/cdc-ansible`
 - Need to be included in the release note